+++ src/intl/en.json
@@ -89,10 +89,10 @@
   },
   "category": {
     "poi": "POI",
-    "catalyst" : "Catalyst",
-    "ban_name" : "Ban Name",
-    "grant" : "Grant",
-    "poll" : "Poll",
+    "catalyst": "Catalyst",
+    "ban_name": "Ban Name",
+    "grant": "Grant",
+    "poll": "Poll",
     "all_title": "All proposals",
     "catalyst_title": "Catalyst Node",
     "catalyst_description": "Add a node to the network of community run servers",
